requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blis==0.7.11
catalogue==2.0.10
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
ChatterBot==1.2.12
chatterbot-corpus==1.2.0
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
constantly==23.10.4
cryptography==46.0.4
cymem==2.0.13
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-api-core==2.29.0
google-api-python-client==2.108.0
google-auth==2.48.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
httplib2==0.31.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
langcodes==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mathparse==0.2.7
mdurl==0.1.2
msgpack==1.1.2
murmurhash==1.0.15
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
packaging==26.0
Pillow==10.1.0
pluggy==1.6.0
preshed==3.0.12
proto-plus==1.27.1
protobuf==6.33.5
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==3.13
redis==5.0.1
regex==2026.1.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.3.2
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.17.0
service-identity==24.2.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
sqlparse==0.5.5
srsly==2.5.2
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
zope.interface==8.2
